--- src/lib/atoms/navLogo.svelte
@@ -1,14 +1,0 @@
-<a href="/"
-	><img
-		src="https://hallostrandeiland.nl/blobs/gebiedsplatform/w250h250/98691/2022/50/hseTekengebied_8.png"
-		alt="logo"
-	/></a
->
-
-<style>
-	img {
-		margin: 12px;
-		width: 70px;
-		height: 70px;
-	}
-</style>
