--- src/lib/atoms/persoonImage.svelte
@@ -1,16 +1,0 @@
-<article class="persoon_foto">
-    <img
-      src="https://i1.wp.com/www.elskerollema.nl/wp-content/uploads/2016/04/elske_rollema.jpeg?w=240&ssl=1"
-      alt="Foto van persoon"
-    />
-</article>
-
-
-<style>
-    .persoon_foto {
-        align-self: center;
-        border: 2px solid var(--color-primary);
-        width: 15.25rem;
-        height: 20.2rem;
-}
-</style>
