--- src/lib/atoms/persoonLinkImage.svelte
@@ -1,17 +1,0 @@
-<div class="linkedin-logo">
-    <a href="https://nl.linkedin.com/in/elske-rollema-920b052">
-      <img
-        src="https://upload.wikimedia.org/wikipedia/commons/c/ca/LinkedIn_logo_initials.png?20140125013055"
-        class="linkedin-image"
-        alt="klikbare linkedin logo"
-      /></a>
-</div>
-
-
-<style>
-    .linkedin-image {
-        margin-top: 1rem;
-        width: 4rem;
-        cursor: pointer;
-}
-</style>
