--- src/lib/atoms/persoonTitle.svelte
@@ -1,12 +1,0 @@
-<article class="#">
-    <h1 class="persoon_naam">Elske Rollema</h1>
-</article>
-
-
-
-<style>
-    .persoon_naam {
-        font-weight: bold;
-        font-size: 26px;
-}
-</style>
