--- src/lib/atoms/risico.svelte
@@ -1,26 +1,0 @@
-<p>Risiconiveau: <img src="https://cdn-icons-png.flaticon.com/512/5318/5318148.png" alt="risicoicoon"> <img src="https://cdn-icons-png.flaticon.com/512/5318/5318148.png" alt="risicoicoon"></p>
-
-<p class="kosten">Kostenindicatie: <span>€€€</span></p>
-
-<style>
-    img{
-        height: 30px;
-        width: 2em;
-        margin-left: 1em;
-    }
-    p{
-        display: flex;
-        flex-flow: row wrap;
-        align-items: center;
-        padding: 0em 0em 0em 1.5em;
-        font-weight: bold;
-    }
-
-    span {
-        padding-left: 1rem;
-        font-size: xx-large;
-        font-weight: lighter;
-        color:var(--color-blue);
-    }
-
-</style>
