+++ src/lib/components/Reactie.svelte
@@ -1,11 +1,10 @@
-<script>
+<!-- <script>
 	export let data;
 	let showImage = true; // Stel in op 'false' als er geen externe afbeelding is
 </script>
 
 <section>
 	<article>
-		<!-- {#each reactions as reaction} -->
 		<figure>
 			<svg width="" height="" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
 				<path
@@ -21,7 +20,6 @@
 			<time datetime="2023-09-27">27 september 2023</time>
 			<p>Dit is een citaat of een stukje tekst dat je wilt benadrukken.</p>
 		</blockquote>
-		<!-- {/each} -->
 	</article>
 </section>
 
@@ -101,4 +99,4 @@
 			margin-left: 20%;
 		}
 	}
-</style>
+</style> -->
