--- src/lib/index.js
@@ -1,9 +1,0 @@
-// place files you want to import through the `$lib` alias in this folder.
-
-// Personen component - Rayan
-
-export { default as Title } from './atoms/persoonTitle.svelte'
-export { default as Image } from './atoms/persoonImage.svelte'
-export { default as LinkImage } from './atoms/persoonLinkImage.svelte'
-
-export { default as TitleBeschrijving } from './molecules/persoonTitleBeschrijving.svelte'
