--- src/routes/personen/+page.svelte
@@ -1,8 +1,0 @@
-<script>
-    import Navbar from '$lib/molecules/navigation.svelte';
-    import Personen from '$lib/organisms/personen.svelte';
-</script>
-
-<Navbar />
-
-<Personen />
