+++ static/styles/global.css
@@ -1,9 +1,10 @@
+
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;600;700,800&display=swap');
 :root {
-    /* Animaties */
-    --animation-default: 0.2s;
 
     /* Eenheden */
-    color: rgb(0, 0, 0);
+    --margin-right: 80px;
+    --margin-left: 80px;
     --unit-deca: 10rem;
     --unit-large: 2rem;
     --unit-default: 1rem;
@@ -13,6 +14,13 @@
     --unit-pico: .5px;
 
     /* Kleuren */
+    --grey: #6B7071;
+    --dark-grey: #333333;
+    --black: #000000;
+    --white: #ffffff;
+    --yellow: #FFE704;
+    --beige: #F0F0E9;
+    --blue: #019CE0;
     --color-primary: #121212;
     --color-primary-50: #495057;
     --color-secundary: #f7f7f7;
@@ -22,29 +30,42 @@
     --color-accent-50: #f0f0eb;
     --color-blue: #0594c6;
 
+    /* Font */
+    --font-family-sans-serif: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, "Noto Sans", "Liberation Sans", sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
+    --font-family-monospace: SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace;
+    --font-weight-200:200;
+    --font-weight-300:300;
+    --font-weight-400:400;
+
     /* Schaduwen  */
     --shadow-default: 0px 2px 8px rgba(135, 135, 135, 0.1), 0px 4px 4px -2px rgba(135, 135, 135, 0.12), 0px 8px 16px -4px rgba(0, 0, 0, 0.2);
-
-}
-
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300&display=swap');
-
-@font-face {
-    font-family: "Poppins";
-    src: url('https://fonts.googleapis.com/css2?family=Poppins:wght@300&display=swap');
 }
 
 * {
-    box-sizing: border-box;
     margin: 0;
     padding: 0;
-}
+    box-sizing: border-box;
 
+}
+*,::before, ::after{
+    box-sizing: border-box;
+}
+html {
+    font-family: sans-serif;
+    line-height: 1.15;
+    -webkit-text-size-adjust: 100%;
+    -webkit-tap-highlight-color:transparent
+}
 body {
-    font-family: 'Poppins', sans-serif;
-    background-color: var(--color-accent-75);
-}
+    margin: 0;
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, "Noto Sans", "Liberation Sans", sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
+    font-size: 1rem;
+    font-weight: var(--font-weight-400);
+    line-height: 1.5;
+    color: var(--dark-grey);
+    text-align: left;
 
+}
 /* Interactief */
 a {
     text-decoration: none;
@@ -90,77 +111,121 @@ mark {
     border: 1px solid var(--color-accent-50);
     color: var(--color-primary-50);
 }
-
-/* dropdown */
-
-.hide {
-    display: none;
-  }
-
-/* Zoekbalk */
-nav figcaption {
-    width: 0;
-    visibility: hidden;
-}
-
-#custom-view {
-    transition: var(--animation-default) ease-in-out;
-}
-
-/* Wensen in rasterweergave */
-.grid-overview article {
-    display: grid;
-    grid-template-columns: 1fr;
-    gap: var(--unit-default);
-}
-
-.grid-overview li {
-    width: 20rem;
-}
-
-/* Wensen in lijstweergave */
-.list-overview article li {
-    display: grid;
-    grid-template-columns: 1fr;
-    align-items: center;
-    gap: var(--unit-large);
-    padding: var(--unit-small);
-    margin-bottom: var(--unit-default);
-    border-bottom: 2px solid var(--color-accent-75);
-}
-
-.list-overview ul li img {
-    width: clamp(8rem, 100%, 10rem);
-}
-
-
-
-@media (min-width: 440px) {
-    .grid-overview article {
-        grid-template-columns: 1fr 1fr;
+/* Responsieve */
+/* Navbar */
+@media screen and (max-width: 800px) {
+    /* nav {
+        position: fixed;
+        z-index: 3;
+    } */
+    
+    .hamburger {
+        display: block;
+        position: absolute;
+        cursor: pointer;
+        right: 5%;
+        top: 50%;
+        transform: translate(-5%, -50%);
+        z-index: 2;
+        transition: all 0.7s ease;
+    }
+    
+    .nav-links {
+        position: fixed;
+        background: var(--beige);
+        height: 50vh;
+        width: 100%;
+        flex-direction: column;
+        clip-path: circle(60px at 90% -40%);
+        -webkit-clip-path: circle(50px at 90% -30%);
+        transition: all 1s ease-out;
+        pointer-events: none;
+    }
+    
+    .nav-links.open {
+        clip-path: circle(1000px at 90% -30%);
+        -webkit-clip-path: circle(1000px at 90% -30%);
+        pointer-events: all;
+    }
+    
+    .nav-links li {
+        opacity: 0;
+    }
+    
+    .nav-links li:nth-child(1) {
+        transition: all 0.5s ease 0.2s;
+    }
+    
+    .nav-links li:nth-child(2) {
+        transition: all 0.5s ease 0.4s;
+    }
+    
+    .nav-links li:nth-child(3) {
+        transition: all 0.5s ease 0.6s;
+    }
+    
+    .nav-links li:nth-child(4) {
+        transition: all 0.5s ease 0.7s;
+    }
+    
+    .nav-links li:nth-child(5) {
+        transition: all 0.5s ease 0.8s;
+    }
+    
+    .nav-links li:nth-child(6) {
+        transition: all 0.5s ease 0.9s;
+        margin: 0;
+    }
+    
+    .nav-links li:nth-child(7) {
+        transition: all 0.5s ease 1s;
+        margin: 0;
+    }
+    
+    .nav-links li:nth-child(8) {
+        transition: all 0.5s ease 1.1s;
+        margin: 0;
+    }
+    
+    li.fade {
+        opacity: 1;
     }
 }
 
-@media (min-width: 600px) {
-    main {
-        width: 50rem;
+/* Mobile */
+@media (max-width: 576px) {
+    .hero-content {
+        padding: 0 10px;
+        flex-direction: column;
     }
 
-    .list-overview article li {
-        grid-template-columns: 12rem 1fr;
-        height: fit-content;
-        width: auto;
-      }
 }
-
-@media (min-width: 628px) {
-    .grid-overview article {
-        grid-template-columns: 1fr 1fr 1fr;
-        margin: auto;
+/* Mobile to tablet */
+@media (max-width: 680px) {
+    .nav-top-links {
+        width: 100%;
+        display: flex;  
+        justify-content: end;
     }
-
-    nav figcaption {
-        width: fit-content;
-        visibility: visible;
+    .hero-content{
+        display: flex;
+        justify-content: space-around;
     }
 }
+/* Footer */
+@media (max-width: 768px) {
+    .footer-content {
+        grid-template-columns: repeat(1, 1fr);
+        row-gap: 20px;
+    }
+    .footer-menu {
+        text-align: left;
+    }
+    .footer-menu li {
+        display: block;
+        margin-bottom: 10px;
+    }
+}
+
+
+
